Gui (#1)
* WIP GUI

Currently has a keypad, a text display, and a
clear button for the text.

* Improve keypad

Resized the keypad to fit the raspberry pi.
Added a period key and a Enter/Return key.

General cleanup aswell.

* Syntax: Clean up some functions

Removed unnecessary stuff.

* Fix: Correct Merge Error

During the merge in db2bee7 a line
was changed which resulted in an
incorrect keypad layout. This commit
fixes that.

* WIP: Scanning cards

Working on the logic for scanning cards. Committing
to allow for remote work.

* Error Check: Add to scanCard

Added error checking to see if the card
reader is plugged into the pi or not.

* Syntax: Create store class

Changed code to be a class.

* Fix: Print Bug

Mended an issue in the printMessage function
so that when a message is made it updates the textbox.

* WIP: Retrieve input from textfield

Along with some general cleanup

* feat: defaultLayout function

This allows for me to redo the layout and hide
buttons for different layers.

* etc: Possible way to fix inputs

* Undo: CreateKeyPad()

Moved back into constructor not in default layout.
I don't want to have to edit the keypad. It's gonna
be static until I have a month to work with this again.

* WIP: Add balance update buttons.

* feat: Create database access

This class creates a database object
which will allow for only usage of stored
procedures that the storeOperator has access to.

* feat: Overhaul nfcScript.py

Removed unnecessary code from file.
Removed all usage of old database code.

* Delete: Remove unnecessary files

Removed old bucky gui. This was originally
for learning to code a gui together.

Removed output. This was used as a debug tool.

* fix: Wrong function call

Fixed subBalButton which called a
nonexistent function.

* fix: Layout bug

the Subtract balance button was not
being turned off on layout switch.

* feat: Change returns

Returns no longer pass in data passed to
the functions for database.py

* WIP: Test new logic

* WIP FIX: Fix import

* WIP: Add account buttons.

* WIP Patch: Change retrieve input to return

* WIP: AddAcc Fix

Slices were not done properly. Now inputs are taken.

* Cleanup: Remove retrieve button

The retrieve button was used during development
and is not needed in the actual code.

* Syntax: Add newline

Missing newline character on scancard method.

* feat Implement balance changers

* bug: remove retrieve

* Bug: Fix buttons on press
Fixed issue with buttons not correctly going back to defaults.